backend/requirements.txt
flask==3.1.3
flask-cors==6.0.2
requests>=2.31.0
python-dotenv>=1.0.1
asgiref==3.11.1